Split routers
